LPS-80392 Move impl of captcha to impl bundle
