+++ modules/apps/foundation/captcha/captcha-impl/build.gradle
@@ -1,4 +1,6 @@
 dependencies {
+	compileInclude group: "com.liferay", name: "nl.captcha.simplecaptcha", version: "1.1.1"
+
 	compileOnly group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
 	compileOnly group: "javax.portlet", name: "portlet-api", version: "2.0"
