+++ modules/apps/foundation/captcha/captcha-impl/src/main/java/com/liferay/captcha/internal/configuration/CaptchaConfigurationModelListener.java
@@ -15,7 +15,7 @@
 package com.liferay.captcha.internal.configuration;
 
 import com.liferay.captcha.configuration.CaptchaConfiguration;
-import com.liferay.captcha.recaptcha.ReCaptchaImpl;
+import com.liferay.captcha.internal.recaptcha.ReCaptchaImpl;
 import com.liferay.portal.configuration.persistence.listener.ConfigurationModelListener;
 import com.liferay.portal.configuration.persistence.listener.ConfigurationModelListenerException;
 import com.liferay.portal.kernel.captcha.CaptchaConfigurationException;
